Prototype notebook for plotting inter-model comparisons (if comma-separated odivs provided)
Parses string of experiment IDs into a list (including of length 1 if a single ID with no commas).
For plotting maps of values or model-observation biases, it loops through all experiment IDs.
If more then 1 ID is provided, it also plots the differences between all of the IDs relative to
the first one in the list (except the first one, which is trivial).